always_cms/requirements.txt
flask == 2.1.1
flask-sqlalchemy == 2.5.1
mysqlclient == 2.1.0
flask-login == 0.6.0
Flask-WTF == 1.0.1
Flask_Babel == 2.0.0
pyotp == 2.6.0
flask-dropzone == 1.6.0
jinja2 == 3.1.1
flask-plugins == 1.6.1
flask-ckeditor == 0.4.6
python-slugify == 6.1.1
Flask-Minify == 0.39
Flask-QRcode == 3.0.0
email-validator == 1.2.1
Pillow == 9.1.1
pip~=21.0.1
boto3==1.24.20